feat(ui-kit): refactor multiple-choice components to multiple-check, update imports and stories
